requirements.txt
python-dotenv
daily-python
pyaudio
requests
numpy
sounddevice
pvporcupine
adafruit-circuitpython-neopixel
adafruit-circuitpython-mpu6050
